Include check for PropagationFlags
The ACL may only apply to subfolders. According to [1], the ACL applies
to the target folder iff InheritOnly is not part of the
PropagationFlags.

There may be an ACL saying we have GenericWrite access to C:\, but if
the InheritOnly flag is set, it only applies to subfolders. To avoid
false positives, we check for it here.

[1] https://docs.microsoft.com/en-us/previous-versions/dotnet/netframework-4.0/ms229747(v=vs.100)?redirectedfrom=MSDN