Adjust Get-ModfifiablePath functions
The ``Get-ModifiablePath`` function of PowerUp.ps1 now uses an xor
operation to compare the ACL of an object to the desired (writable)
access mask. This should be slightly faster than comparing the string
values of the ACL.

Furthermore, the error handling was improved (well, before it was
basically not existing) and the Ownership of an object was added as a
property to the returned object.

The new ``Get-ModifiablePath`` function was also added to a seperate
script PowerUpLight.ps1.