Add C:\ to the blacklist of Get-UnqoutedServiceWmi
The ``C:\`` directory is always modifiable for low privileged user
accounts, as they are able to create directories within it. This fact is
pretty annoying for PowerUp invocations, as for the folders ``C:\Program
Files`` and ``C:\Program Files(x86)``, PowerUp will always check if
``C:\`` is writable for Unquoted-Service checks. This creates almost
always false positives and therefore ``C:\`` was moved to a blacklist.