Allow PowerUp.Service objects for Tets-ServiceDacl
The ``Test-ServiceDaclPermission`` cmdlet of PowerUp used only service
names for permission enumeration. The service name was used to get a
service handle and from this handle, the permissions were enumerated.

With PowerUp.Service objects, this approach is superflous, as the
permissions are already assigned by reading the security key from the
registry key. Therfore, an additional check was added and if the
incoming object to Test-ServiceDaclPermission does already has a dacl
property, adding the dacl is skipped.