Remove PowerUp WMI functions
During all changes so far, I kept most of the original PowerUp
functionality. However, I decided now that some of the functions are
just obsulete and it makes no sense to keep them longer.

To be more detailed: The PowerUp Service checks did all start with a
call to enumerate all available services. This is redundant, as the
service objects can be obtained once and then used inside of the
functions.

In furture commits, PowerUp will obtain a set of service objects once
(either by using the resgirty, Get-Service or WMI) and work with it.