Make it work :)
Finally we are back at a working version of PowerUp.ps1. The
Invoke-AllChecks methods should work now again and this even more
reliable than before.

Furthermore, the warning messages have changed in this commit. I noticed
that it is far better to use ``Write-Verbose`` instead of
``Write-Warning`` for PowerUp checks, as this enables you to use the
``-Verbose`` switch of ``Invoke-AllChecks``.