Add second function to check for mod. schtask file
PowerUp does already include a check for modifiable executable paths on
scheduled tasks. However, this check relies on read access to the
``C:\Windows\System32\Tasks`` folder, which is not granted by default on
modern systems. Therefore, an additional method was added that tries to
enumerate scheduled tasks using the schtasks.exe executable and checks
for modifiable paths.