Add owner check to Get-ModifiablePath
So far, ``Get-ModifiablePath`` did only check for explicit permissions
to modify the object. However, in rare situations it could be, that the
user does not has permissions set, but is the owner of the file in
question. In these cases the user has implicit FullControl, which should
also be reported.