Improve Get-ModifiableServiceFile
The Get-ModifiableServiceFile check did only check if the service binary
is writable. However, there are other options for vulnerable services,
e.g. if one of the containing directories is writable or under full
controll of low privileged user accounts. Other common cases are that
the binary itself is not writable, but shared libraries in a parent
folder are. All this stuff would be missed by powerup so far.

For sure, one can not catch all of the above mentioned issues, but with
the applied catch from this commit, PowerUp does at least check the
parent folders of the service executables. Furthermore, it checks
whether a service executable exists. Non existing service executables
are always something, one should take note of.